fix new clang-tidy warnings: performance-enum-size
https://clang.llvm.org/extra/clang-tidy/checks/performance/enum-size.html

+++ headers/meta.hpp/meta_detail/type_kinds.hpp
@@ -10,7 +10,7 @@
 
 namespace meta_hpp::detail
 {
-    enum class type_kind : std::uint32_t {
+    enum class type_kind : std::uint8_t {
         array_,
         class_,
         constructor_,
